fish/requirements.txt
aioredis==1.3.1
amqp

asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn
Automat==20.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.2
certifi==2021.5.30

channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.4
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.6
django-admin-view-permission==1.9
django-background-task==0.1.8
django-background-tasks==1.1.13
django-compat==1.0.15
django-cors-headers==3.7.0
django-mathfilters==1.0.0
django-sslserver==0.22
django-utils-six==2.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
et-xmlfile==1.1.0
flower==1.0.0
hazm==0.7.0
hiredis
humanize==3.13.1
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.11.4
incremental==21.3.0
jdatetime==3.6.2
kombu==5.2.2
msgpack==1.0.4
mysqlclient
nltk==3.3
openpyxl==3.0.10
Pillow
prometheus-client
prompt-toolkit==3.0.24
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyfcm==1.5.3
pyOpenSSL==22.0.0
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.26.0
schedule==1.1.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
tornado==6.1
Twisted

txaio==22.2.1
typing_extensions==4.2.0
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.8.0
zope.interface==5.4.0
